use named index, return index, move helpers to own file
